backend/requirements.txt
Django>=6.0.0
djangorestframework>=3.16.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai>=0.8.0
protobuf>=5.0.0,<6.0.0
Pillow>=10.3.0

